requirements.txt
google-cloud-documentai==2.21.1
python-dotenv==1.0.0
